requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
coverage==6.3.2
Django==4.0.3
djangorestframework==3.13.1
pytz==2022.1
sqlparse==0.4.2
